financeService created (Read Commit Notes)
+ [POST] - createArrearsByGrade = creates Arrears by Grade and Academic Year. Useful if all students in a grade study the same study courses; and these students are enrolled in the same academic year.

+ [POST] - createArrearsByAcademicYear = creates Arrears by Grade and Academic Year. Useful if all students are in the same academic year and the arrear applies to all students of that year. E.g sports_fee, ECA charges etc.

+ [GET] - getStudentFee = Returns the student fee by ID. Fee is sum of all arrears.

Arrear Definition => An arrear is a financial entity of any user. It has an amount that can be positive or negative. If the amount is positive, it means that the user owes the institution, if the amount is negative, it means that the user paid back to the institution.

Working Example:

Tuition Fall 2023 : 24000 PKR
Parent Pays Fee : -23999 PKR
Student Fee : 1 PKR