createdEnrollmentService, Finished academicYearService and major Database changes
Major DB Change!
--
Elephant in the room. ClassID. Having numeric classIDs is useless. Imagine you've been running your school for some 10 years. Class 10B of 2024 has classID 127. The number 127 has no relation with the grade, section or year.

For an admin member to lookup the students of this class. He needs to first search the classID of 10B of 2024 and then get the students.

The new classID is:
classID = grade-section-academicYear

This makes lookup more intuitive and negates the searching of a classID.

#Learn2Code
(Also whose bright fucking idea was to prefix classID with a 'C', unnecessary string processing when a simple integer can be passed. We should do it when we have a unified search function which searches for class, course, faculty, student and guardians and we need to clarify what type of thing we want to search.)

Due to this change, tables would need to be dropped and remade. The correct procedure is stored in the SQL File. Please paste the SQL code after the reference commit comment.

ClassID is now VARCHAR(9) instead of INTEGER AUTO_INCREMENT. Furthermore, most Class APIs have been tested and made compatible with this change. The same goes for student!
--

Minor DB Changes
--
+ AcademicYear now has a boolean Flag.
+ Student no longer has a classID. A seperate table called studentAcademicHistory contains the student's academic history.
+ Class > Section is now VARCHAR(2)
--

APIs
--
+ [POST] createAcademicYear - academicYearService
+ [PATCH] setAcademicYearDays - academicYearService
+ [PATCH] deleteAcademicYear - academicYearService
+ [PATCH] restoreAcademicYear - academicYearService
+ [GET] getAcademicYears - academicYearService
+ [GET] getAcademicYear - academicYearService
+ [POST] enrollStudentToClass - enrollmentService
+ [POST] denrollStudentFromClass - enrollmentService
+ [GET] getClassByID - classService
+ [GET] getClassByAcademicYear - classService
+ [GET] getStudentsByClass - enrollmentService
+ [GET] getStudentsByClassID - enrollmentService
+ and more...(I lost track)

Fixes
--
+ Added getByClassID test
+ Moved createAcademicYear to academicYearService
+ An Unknown Number of APIs have been fixed/created